tests/wpt/tests/css/css-flexbox/parsing/flex-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Flexible Box Layout: getComputedStyle().flex</title>
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#propdef-flex">
<meta name="assert" content="flex computed value is specified keywords and/or computed <length-percentage>.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #container {
    container-type: inline-size;
    width: 10px;
  }
  #target {
    font-size: 40px;
  }
</style>
</head>
<body>
<div id="container">
    <div id="target"></div>
</div>
<script>
test_computed_value("flex", "none", "0 0 auto");
test_computed_value("flex", "1", "1 1 0%");
test_computed_value("flex", "2 3", "2 3 0%");
test_computed_value("flex", "4 5 6px");
test_computed_value("flex", "7% 8", "8 1 7%");
test_computed_value("flex", "8 auto", "8 1 auto");
test_computed_value("flex", "calc(10px + 0.5em)", "1 1 30px");
test_computed_value("flex", "calc(10px - 0.5em)", "1 1 0px");
test_computed_value("flex", "1 1 calc(10em)", "1 1 400px");
test_computed_value("flex", "1 1 calc(-10em)", "1 1 0px");
test_computed_value("flex", "calc(10 + (sign(20cqw - 10px) * 5)) calc(10 + (sign(20cqw - 10px) * 5)) 1px", "5 5 1px");
test_computed_value("flex", "1 1 calc(10px + (sign(20cqw - 10px) * 5px))", "1 1 5px");
test_computed_value("flex", "calc(1) calc(2 + 1) calc(3px)", "1 3 3px");
test_computed_value("flex", "calc(-1) calc(-1) 0", "0 0 0px");
</script>
</body>
</html>
