third_party/blink/web_tests/external/wpt/css/css-flexbox/parsing/flex-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property flex value 'calc(10 + (sign(20cqw - 10px) * 5)) calc(10 + (sign(20cqw - 10px) * 5)) 1px'
  assert_equals: expected "5 5 1px" but got "15 15 1px"
[FAIL] Property flex value 'calc(-1) calc(-1) 0'
  assert_true: 'calc(-1) calc(-1) 0' is a supported value for flex. expected true got false
Harness: the test ran to completion.
